+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Change some `java.util.List` to `java.util.Collection` in `AbstractSessionContextImpl`
 * Change `ReplicantSession.isSubscriptionEntryPresent()` to be public
+* Add `ReplicantSession.recordEntityScopedGraphLink(...)` for downstream post-subscribe entity-scoped graph links.
 
 ### [v6.189](https://github.com/replicant4j/replicant/tree/v6.189) (2026-05-08) · [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.188...v6.189)